feat: add settings for Input.Preview
